bq2597x_charger: Modify floating point operation in bq2597x_get_adc_data
ld.lld: error: undefined symbol: __floatsidf
>>> referenced by bq2597x_charger.c:0 (..out/target/product/tucana/obj/KERNEL_OBJ/../../../../../../kernel/xiaomi/tucana/drivers/power/supply/ti/bq2597x_charger.c:0)
>>>               vmlinux.o:(bq2597x_get_adc_data)

ld.lld: error: undefined symbol: __muldf3
>>> referenced by bq2597x_charger.c:0 (../out/target/product/tucana/obj/KERNEL_OBJ/../../../../../../kernel/xiaomi/tucana/drivers/power/supply/ti/bq2597x_charger.c:0)
>>>               vmlinux.o:(bq2597x_get_adc_data)

ld.lld: error: undefined symbol: __fixdfsi
>>> referenced by bq2597x_charger.c:0 (..out/target/product/tucana/obj/KERNEL_OBJ/../../../../../../kernel/xiaomi/tucana/drivers/power/supply/ti/bq2597x_charger.c:0)
>>>               vmlinux.o:(bq2597x_get_adc_data)

+++ drivers/power/supply/ti/bq2597x_charger.c
@@ -68,7 +68,7 @@ enum {
 	ADC_MAX_NUM,
 };
 
-static float sc8551_adc_lsb[] = {
+static long sc8551_adc_lsb[] = {
 	[ADC_IBUS]	= SC8551_IBUS_ADC_LSB,
 	[ADC_VBUS]	= SC8551_VBUS_ADC_LSB,
 	[ADC_VAC]	= SC8551_VAC_ADC_LSB,
@@ -80,7 +80,7 @@ static float sc8551_adc_lsb[] = {
 	[ADC_TDIE]	= SC8551_TDIE_ADC_LSB,
 };
 
-static float sc8551_adc_lsb_non_calibrate[] = {
+static long sc8551_adc_lsb_non_calibrate[] = {
 	[ADC_IBUS]	= SC8551_IBUS_ADC_LSB,
 	[ADC_VBUS]	= SC8551_VBUS_ADC_LSB,
 	[ADC_VAC]	= SC8551_VAC_ADC_LSB,
